# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PharmaCare Digital Pharmacy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conditions/{id}:
    get:
      operationId: getCondition
      tags: [conditions]
      summary: Get condition details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Condition details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condition"
        "404":
          description: Not found
...
  /consultations/{id}/messages:
    get:
      operationId: listConsultationMessages
      tags: [consultations]
      summary: List chat messages on a consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConsultationMessage"
                  actions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConsultationAction"
                required: [messages, actions]
    post:
      operationId: postConsultationMessage
      tags: [consultations]
      summary: Post a chat message on a consultation (patient or pharmacist)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationMessage"
...
  /auth/pharmacist-login:
    post:
      operationId: pharmacistLogin
      tags: [auth]
      summary: Pharmacist login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PharmacistLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PharmacistAuthResponse"
        "401":
          description: Invalid credentials
  /auth/patient-register:
    post:
      operationId: patientRegister
      tags: [auth]
      summary: Patient account registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientAuthResponse"
        "409":
          description: Email already registered
  /auth/patient-login:
    post:
      operationId: patientLogin
      tags: [auth]
      summary: Patient login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientAuthResponse"
        "401":
          description: Invalid credentials
...
components:
  schemas:
...
    Condition:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [skin, womens_health, eye_care, digestive, children_family, pain_minor_illness, allergy, respiratory, ear, urinary, weight_management, sexual_performance, sexual_health, stis, hair_skin_nails, seasonal_viruses, travel_health]
        description:
          type: string
        onlineEligible:
          type: boolean
        requiresPhoto:
          type: boolean
        requiresInPerson:
          type: boolean
        ageRestrictions:
          type: string
          nullable: true
        redFlags:
          type: array
          items:
            type: string
        questionsCount:
          type: integer
      required:
        - id
        - name
        - category
        - description
        - onlineEligible
        - requiresPhoto
        - requiresInPerson
        - redFlags
        - questionsCount
...
    ConsultationMessage:
      type: object
      properties:
        id:
          type: string
        consultationId:
          type: string
        patientEmail:
          type: string
        senderRole:
          type: string
        senderName:
          type: string
        body:
          type: string
        kind:
          type: string
        meta:
          type: string
          nullable: true
        readByPatient:
          type: boolean
        readByPharmacist:
          type: boolean
        createdAt:
          type: string
      required: [id, consultationId, patientEmail, senderRole, senderName, body, kind, readByPatient, readByPharmacist, createdAt]
    ConsultationMessageInput:
      type: object
      properties:
        body:
          type: string
        kind:
          type: string
          nullable: true
      required: [body]
    ConsultationAction:
      type: object
      properties:
        id:
          type: string
        consultationId:
          type: string
        action:
          type: string
        actorRole:
          type: string
        actorName:
          type: string
        details:
          type: object
          additionalProperties: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, consultationId, action, actorRole, actorName, createdAt]
...
    PharmacistLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    PharmacistAuthResponse:
      type: object
      properties:
        token:
          type: string
        pharmacistName:
          type: string
        pharmacistId:
          type: string
        role:
          type: string
      required:
        - token
        - pharmacistName
        - pharmacistId
        - role
    PatientRegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    PatientLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    PatientAuthResponse:
      type: object
      properties:
        token:
          type: string
        patientId:
          type: string
        name:
          type: string
        email:
          type: string
      required:
        - token
        - patientId
        - name
        - email
...